# backend_school_crm/docs/swagger.yaml
basePath: /api/v1
definitions:
  models.CreateSubscriptionRequest:
    properties:
      branch_id:
        type: string
      notes:
        type: string
      payment_method:
        type: string
      plan_id:
        type: string
    required:
    - plan_id
    type: object
...
  models.Subscription:
    properties:
      auto_renew:
        type: boolean
      branch_id:
        type: string
      cancelled_at:
        type: string
      cancelled_by:
        type: string
      created_at:
        type: string
      end_date:
        type: string
      id:
        type: string
      notes:
        type: string
      payment_method:
        type: string
      plan_id:
        type: string
      renewal_date:
        type: string
      start_date:
        type: string
      status:
        description: active, paused, cancelled, expired
        type: string
      stripe_subscription_id:
        type: string
      updated_at:
        type: string
      user_id:
        type: string
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@school-crm.example
    name: School CRM Support
  description: Residual monolith — auth, subscriptions, Click.uz/Telegram payments,
    messaging, reports, dev panel.
  license:
    name: Proprietary
  termsOfService: http://swagger.io/terms/
  title: School CRM API
  version: "1.0"
paths:
...
  /expenses:
    get:
      description: Returns expense records for a branch, optionally filtered by date
        range.
      parameters:
      - description: Branch ID
        in: query
        name: branchId
        required: true
        type: string
      - description: YYYY-MM-DD
        in: query
        name: startDate
        type: string
      - description: YYYY-MM-DD
        in: query
        name: endDate
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: List expenses
      tags:
      - expenses
    post:
      consumes:
      - application/json
      parameters:
      - description: title, amount, category, branch_id, date, payment_method
        in: body
        name: body
        required: true
        schema:
          additionalProperties: true
          type: object
      produces:
      - application/json
      responses:
        "201":
          description: created expense
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Create expense
      tags:
      - expenses
  /expenses/{id}:
    delete:
      parameters:
      - description: Expense UUID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: expense deleted
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Delete expense
      tags:
      - expenses
...
  /payments:
    get:
      description: Returns payments for a branch, with optional month/year/status
        filters.
      parameters:
      - description: Branch ID
        in: query
        name: branchId
        required: true
        type: string
      - description: Month (1-12)
        in: query
        name: month
        type: integer
      - description: Year (e.g. 2025)
        in: query
        name: year
        type: integer
      - description: paid | partial | not_paid
        in: query
        name: status
        type: string
      - description: Page
        in: query
        name: page
        type: integer
      - description: Limit
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: List payments
      tags:
      - payments
    post:
      consumes:
      - application/json
      description: Records a new payment entry for a student.
      parameters:
      - description: student_id, branch_id, amount, month, year, payment_method
        in: body
        name: body
        required: true
        schema:
          additionalProperties: true
          type: object
      produces:
      - application/json
      responses:
        "201":
          description: created payment
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Create payment
      tags:
      - payments
  /payments/{id}:
    delete:
      description: Permanently removes a payment record.
      parameters:
      - description: Payment UUID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: payment deleted
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Delete payment
      tags:
      - payments
    put:
      consumes:
      - application/json
      parameters:
      - description: Payment UUID
        in: path
        name: id
        required: true
        type: string
      - description: Fields to update
        in: body
        name: body
        required: true
        schema:
          additionalProperties: true
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Update payment
      tags:
      - payments
...
  /subscriptions:
    post:
      consumes:
      - application/json
      description: Create a new subscription for the current user
      parameters:
      - description: Create subscription request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.CreateSubscriptionRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.Subscription'
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: Create subscription
      tags:
      - subscriptions
...
securityDefinitions:
  BearerAuth:
    description: JWT token — prefix with "Bearer "
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"